+++ activejob/test/cases/adapter_test.rb
@@ -33,4 +33,20 @@ class AdapterTest < ActiveSupport::TestCase
       end
     end
   end
+
+  if adapter_is?(:sidekiq)
+    test "sidekiq adapter should be deprecated" do
+      before_adapter = ActiveJob::Base.queue_adapter
+
+      msg = <<~MSG.squish
+        The built-in `sidekiq` adapter is deprecated and will be removed in Rails 8.1.
+        Please upgrade `sidekiq` gem to version 7.3.3 or later to use the `sidekiq` gem's adapter.
+      MSG
+      assert_deprecated(msg, ActiveJob.deprecator) do
+        ActiveJob::Base.queue_adapter = :sidekiq
+      end
+    ensure
+      ActiveJob::Base.queue_adapter = before_adapter
+    end
+  end
 end