+++ activejob/CHANGELOG.md
@@ -1,2 +1,9 @@
+*   Deprecate built-in `sidekiq` adapter.
+
+    If you're using this adapter, upgrade to `sidekiq` 7.3.3 or later to use the `sidekiq` gem's adapter.
+
+    *fatkodima*
+
+## Rails 8.0.0.beta1 (September 26, 2024) ##
 
 Please check [8-0-stable](https://github.com/rails/rails/blob/8-0-stable/activejob/CHANGELOG.md) for previous changes.